@@ -1,15 +1,15 @@
 [build-system]
 requires = [
-    "setuptools>=77.0.0",
-    "jax >= 0.4.31; python_version > '3.9' and (sys_platform == 'win32' or sys_platform == 'linux')"
+    "scikit-build-core>=0.10",
+    "jax >= 0.4.31; sys_platform == 'win32' or sys_platform == 'linux'"
 ]
-build-backend = "setuptools.build_meta"
+build-backend = "scikit_build_core.build"
 
 [project]
 name = "ale-py"
 description = "The Arcade Learning Environment (ALE) - a platform for AI research."
 readme = "README.md"
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 license = "GPL-2.0-only"
 license-files = ["LICENSE.md"]
 keywords = ["reinforcement-learning", "arcade-learning-environment", "atari"]
